+++ tests/unit/fixtures.py
@@ -1,0 +1,8 @@
+import pytest
+from syrupy.assertion import SnapshotAssertion
+
+
+class ProgramSnapshot:
+    @pytest.fixture(autouse=True)
+    def _setup(self, snapshot_xml: SnapshotAssertion):
+        self.snapshot_xml = snapshot_xml
